Split settings classes up into separate files
